Break out sortItems in App.js
